backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pydantic
pydantic[email]
apscheduler
aiosmtplib
openai
pytz
imap-tools
beautifulsoup4
httpx